+++ eng/common/scripts/Update-ChangeLog.ps1
@@ -89,8 +89,7 @@ LogDebug "The latest release note entry in the changelog is for version [$($Late
 
 $LatestsSorted = [AzureEngSemanticVersion]::SortVersionStrings(@($LatestVersion, $Version))
 if ($LatestsSorted[0] -ne $Version) {
-    LogWarning "Version [$Version] is older than the latestversion [$LatestVersion] in the changelog. Please use a more recent version."
-    exit(0)
+    LogWarning "Version [$Version] is older than the latestversion [$LatestVersion] in the changelog. Consider using a more recent version."
 }
 
 if ($ReplaceLatestEntryTitle) 
@@ -99,7 +98,7 @@ if ($ReplaceLatestEntryTitle)
     LogDebug "Resetting latest entry title to [$($newChangeLogEntry.ReleaseTitle)]"
     $ChangeLogEntries.Remove($LatestVersion)
     if ($newChangeLogEntry) {
-        $ChangeLogEntries[$Version] = $newChangeLogEntry
+        $ChangeLogEntries.Insert(0, $Version, $newChangeLogEntry)
     }
     else {
         LogError "Failed to create new changelog entry"
@@ -117,7 +116,7 @@ else
     LogDebug "Adding new ChangeLog entry for Version [$Version]"
     $newChangeLogEntry = New-ChangeLogEntry -Version $Version -Status $ReleaseStatus
     if ($newChangeLogEntry) {
-        $ChangeLogEntries[$Version] = $newChangeLogEntry
+        $ChangeLogEntries.Insert(0, $Version, $newChangeLogEntry)
     }
     else {
         LogError "Failed to create new changelog entry"