+++ sdk/communication/Azure.Communication.Administration/src/PhoneNumberAdministrationClient.cs
@@ -41,6 +41,19 @@ namespace Azure.Communication.Administration
                   ConnectionString.Parse(AssertNotNull(connectionString, nameof(connectionString))))
         { }
 
+        /// <summary>
+        /// Initializes a phone number administration client with a token credential.
+        /// <param name="endpoint">The URI of the Azure Communication Services resource.</param>
+        /// <param name="tokenCredential">The <see cref="TokenCredential"/> used to authenticate requests, such as DefaultAzureCredential.</param>
+        /// <param name="options">Client option exposing <see cref="ClientOptions.Diagnostics"/>, <see cref="ClientOptions.Retry"/>, <see cref="ClientOptions.Transport"/>, etc.</param>
+        /// </summary>
+        public PhoneNumberAdministrationClient(Uri endpoint, TokenCredential tokenCredential, PhoneNumberAdministrationClientOptions? options = default)
+            : this(
+                  endpoint,
+                  options ?? new PhoneNumberAdministrationClientOptions(),
+                  tokenCredential)
+        { }
+
         internal PhoneNumberAdministrationClient(PhoneNumberAdministrationClientOptions options, ConnectionString connectionString)
             : this(new ClientDiagnostics(options), options.BuildHttpPipeline(connectionString), connectionString.GetRequired("endpoint"))
         { }
@@ -57,6 +70,18 @@ namespace Azure.Communication.Administration
             RestClient = new PhoneNumberAdministrationRestClient(
                 ClientDiagnostics,
                 options.BuildHttpPipeline(credential),
+                endpoint.AbsoluteUri);
+        }
+
+        internal PhoneNumberAdministrationClient(Uri endpoint, PhoneNumberAdministrationClientOptions options, TokenCredential tokenCredential)
+        {
+            Argument.AssertNotNull(endpoint, nameof(endpoint));
+            Argument.AssertNotNull(tokenCredential, nameof(tokenCredential));
+
+            ClientDiagnostics = new ClientDiagnostics(options);
+            RestClient = new PhoneNumberAdministrationRestClient(
+                ClientDiagnostics,
+                options.BuildHttpPipeline(tokenCredential),
                 endpoint.AbsoluteUri);
         }
 
@@ -301,7 +326,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                return await RestClient.GetNumberConfigurationAsync(phoneNumber.Value, cancellationToken).ConfigureAwait(false);
+                return await RestClient.GetNumberConfigurationAsync(phoneNumber.PhoneNumber, cancellationToken).ConfigureAwait(false);
             }
             catch (Exception ex)
             {
@@ -320,7 +345,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                return RestClient.GetNumberConfiguration(phoneNumber.Value, cancellationToken);
+                return RestClient.GetNumberConfiguration(phoneNumber.PhoneNumber, cancellationToken);
             }
             catch (Exception ex)
             {
@@ -340,7 +365,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                return await RestClient.ConfigureNumberAsync(pstnConfiguration, phoneNumber.Value, cancellationToken).ConfigureAwait(false);
+                return await RestClient.ConfigureNumberAsync(pstnConfiguration, phoneNumber.PhoneNumber, cancellationToken).ConfigureAwait(false);
             }
             catch (Exception ex)
             {
@@ -360,7 +385,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                return RestClient.ConfigureNumber(pstnConfiguration, phoneNumber.Value, cancellationToken);
+                return RestClient.ConfigureNumber(pstnConfiguration, phoneNumber.PhoneNumber, cancellationToken);
             }
             catch (Exception ex)
             {
@@ -379,7 +404,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                return await RestClient.UnconfigureNumberAsync(phoneNumber.Value, cancellationToken).ConfigureAwait(false);
+                return await RestClient.UnconfigureNumberAsync(phoneNumber.PhoneNumber, cancellationToken).ConfigureAwait(false);
             }
             catch (Exception ex)
             {
@@ -398,7 +423,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                return RestClient.UnconfigureNumber(phoneNumber.Value, cancellationToken);
+                return RestClient.UnconfigureNumber(phoneNumber.PhoneNumber, cancellationToken);
             }
             catch (Exception ex)
             {
@@ -630,7 +655,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                var response = await RestClient.ReleasePhoneNumbersAsync(phoneNumbers.Select(phoneNumber => phoneNumber.Value), cancellationToken).ConfigureAwait(false);
+                var response = await RestClient.ReleasePhoneNumbersAsync(phoneNumbers.Select(phoneNumber => phoneNumber.PhoneNumber), cancellationToken).ConfigureAwait(false);
                 return new ReleasePhoneNumberOperation(
                        this,
                        response.Value.ReleaseId,
@@ -673,7 +698,7 @@ namespace Azure.Communication.Administration
             scope.Start();
             try
             {
-                var response = RestClient.ReleasePhoneNumbers(phoneNumbers.Select(phoneNumber => phoneNumber.Value), cancellationToken);
+                var response = RestClient.ReleasePhoneNumbers(phoneNumbers.Select(phoneNumber => phoneNumber.PhoneNumber), cancellationToken);
                 return new ReleasePhoneNumberOperation(
                        this,
                        response.Value.ReleaseId,