+++ sdk/eventhub/Microsoft.Azure.WebJobs.Extensions.EventHubs/src/Config/EventHubWebJobsBuilderExtensions.cs
@@ -42,17 +42,21 @@ namespace Microsoft.Extensions.Hosting
                 {
                     // Map old property names for backwards compatibility
                     // do it before the binding so new property names take precedence
-                    options.InvokeProcessorAfterReceiveTimeout = section.GetValue(
+                    options.InvokeFunctionAfterReceiveTimeout = section.GetValue(
                         "EventProcessorOptions:InvokeProcessorAfterReceiveTimeout",
-                        options.InvokeProcessorAfterReceiveTimeout);
+                        options.InvokeFunctionAfterReceiveTimeout);
 
-                    options.EventProcessorOptions.TrackLastEnqueuedEventProperties = section.GetValue(
+                    options.TrackLastEnqueuedEventProperties = section.GetValue(
                         "EventProcessorOptions:EnableReceiverRuntimeMetric",
-                        options.EventProcessorOptions.TrackLastEnqueuedEventProperties);
+                        options.TrackLastEnqueuedEventProperties);
 
                     options.MaxBatchSize = section.GetValue(
                         "EventProcessorOptions:MaxBatchSize",
                         options.MaxBatchSize);
+
+                    options.PrefetchCount = section.GetValue(
+                        "EventProcessorOptions:PrefetchCount",
+                        options.PrefetchCount);
 
                     var receiveTimeout = section.GetValue<TimeSpan?>(
                         "EventProcessorOptions:ReceiveTimeout",
@@ -60,7 +64,7 @@ namespace Microsoft.Extensions.Hosting
 
                     if (receiveTimeout != null)
                     {
-                        options.EventProcessorOptions.MaximumWaitTime = receiveTimeout.Value;
+                        options.MaximumWaitTime = receiveTimeout.Value;
                     }
 
                     var leaseDuration = section.GetValue<TimeSpan?>(
@@ -69,7 +73,7 @@ namespace Microsoft.Extensions.Hosting
 
                     if (leaseDuration != null)
                     {
-                        options.EventProcessorOptions.PartitionOwnershipExpirationInterval = leaseDuration.Value;
+                        options.PartitionOwnershipExpirationInterval = leaseDuration.Value;
                     }
 
                     var renewInterval = section.GetValue<TimeSpan?>(
@@ -78,7 +82,7 @@ namespace Microsoft.Extensions.Hosting
 
                     if (renewInterval != null)
                     {
-                        options.EventProcessorOptions.LoadBalancingUpdateInterval = renewInterval.Value;
+                        options.LoadBalancingUpdateInterval = renewInterval.Value;
                     }
                 })
                 .BindOptions<EventHubOptions>();