+++ sdk/storage/Azure.Storage.Common/src/Shared/Constants.cs
@@ -88,6 +88,12 @@ namespace Azure.Storage
         public const string FalseName = "false";
         public const string TrueName = "true";
 
+        public const string ErrorCode = "Code";
+        public const string ErrorMessage = "Message";
+
+        public const string CommaString = ",";
+        public const char CommaChar = ',';
+
         /// <summary>
         /// Storage Connection String constant values.
         /// </summary>
@@ -152,6 +158,7 @@ namespace Azure.Storage
             public const string Range = "Range";
             public const string ContentRange = "Content-Range";
             public const string VersionId = "x-ms-version-id";
+            public const string LeaseTime = "x-ms-lease-time";
         }
 
         internal static class ErrorCodes
@@ -343,6 +350,8 @@ namespace Azure.Storage
             public const string MessagesUri = "messages";
 
             public const string UriSubDomain = "queue";
+
+            public const string QueueTraitsMetadata = "metadata";
         }
 
         /// <summary>
@@ -591,6 +600,7 @@ namespace Azure.Storage
         internal static class HttpStatusCode
         {
             internal const int NotFound = 404;
+            internal const int NotModified = 304;
         }
     }
 }