+++ sdk/storage/generate.ps1
@@ -8,12 +8,6 @@ autorest --use=$PSScriptRoot/Azure.Storage.Common/swagger/Generator/ --verbose
 cd $PSScriptRoot/Azure.Storage.Blobs.Batch/swagger/
 autorest --use=$PSScriptRoot/Azure.Storage.Common/swagger/Generator/ --verbose
 
-cd $PSScriptRoot/Azure.Storage.Files.Shares/swagger/
-autorest --use=$PSScriptRoot/Azure.Storage.Common/swagger/Generator/ --verbose
-
-cd $PSScriptRoot/Azure.Storage.Queues/swagger/
-autorest --use=$PSScriptRoot/Azure.Storage.Common/swagger/Generator/ --verbose
-
 cd $PSScriptRoot/Azure.Storage.Files.DataLake/swagger/
 autorest --use=$PSScriptRoot/Azure.Storage.Common/swagger/Generator/ --verbose