+++ sdk/synapse/Azure.Analytics.Synapse.Spark/tests/SparkTestUtilities.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 using System;
@@ -6,27 +6,29 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Azure.Analytics.Synapse.Spark.Models;
 using Azure.Analytics.Synapse.Tests;
+using Azure.Core.TestFramework;
+using Azure.Analytics.Synapse.Spark;
 
-namespace Azure.Analytics.Synapse.Spark.Tests
+namespace Azure.Analytics.Synapse.Tests
 {
-    internal static class SparkTestUtilities
+    public static class SparkTestUtilities
     {
         /// <summary>
         /// Create parameters for Spark batch tests.
         /// </summary>
         /// <param name="test"></param>
         /// <returns></returns>
-        internal static SparkBatchJobOptions CreateSparkJobRequestParameters(this SparkClientTestBase test)
+        public static SparkBatchJobOptions CreateSparkJobRequestParameters(TestRecording recording, SynapseTestEnvironment testEnvironment)
         {
-            string name = test.Recording.GenerateName("dontnetbatch");
-            string file = string.Format("abfss://{0}@{1}.dfs.core.windows.net/samples/java/wordcount/wordcount.jar", test.TestEnvironment.StorageFileSystemName, test.TestEnvironment.StorageAccountName);
+            string name = recording.GenerateId("dontnetbatch", 16);
+            string file = string.Format("abfss://{0}@{1}.dfs.core.windows.net/samples/java/wordcount/wordcount.jar", testEnvironment.StorageFileSystemName, testEnvironment.StorageAccountName);
             return new SparkBatchJobOptions(name, file)
             {
                 ClassName = "WordCount",
                 Arguments =
                 {
-                    string.Format("abfss://{0}@{1}.dfs.core.windows.net/samples/java/wordcount/shakespeare.txt", test.TestEnvironment.StorageFileSystemName, test.TestEnvironment.StorageAccountName),
-                    string.Format("abfss://{0}@{1}.dfs.core.windows.net/samples/java/wordcount/result/", test.TestEnvironment.StorageFileSystemName, test.TestEnvironment.StorageAccountName),
+                    string.Format("abfss://{0}@{1}.dfs.core.windows.net/samples/java/wordcount/shakespeare.txt", testEnvironment.StorageFileSystemName, testEnvironment.StorageAccountName),
+                    string.Format("abfss://{0}@{1}.dfs.core.windows.net/samples/java/wordcount/result/", testEnvironment.StorageFileSystemName, testEnvironment.StorageAccountName),
                 },
                 DriverMemory = "28g",
                 DriverCores = 4,
@@ -41,9 +43,9 @@ namespace Azure.Analytics.Synapse.Spark.Tests
         /// </summary>
         /// <param name="test"></param>
         /// <returns></returns>
-        internal static SparkSessionOptions CreateSparkSessionRequestParameters(this SparkClientTestBase test)
+        public static SparkSessionOptions CreateSparkSessionRequestParameters(TestRecording recording)
         {
-            string name = test.Recording.GenerateName("dotnetsession");
+            string name = recording.GenerateId("dotnetsession", 16);
             return new SparkSessionOptions(name)
             {
                 DriverMemory = "28g",
@@ -54,7 +56,7 @@ namespace Azure.Analytics.Synapse.Spark.Tests
             };
         }
 
-        internal static async Task<List<SparkBatchJob>> ListSparkBatchJobsAsync(this SparkClientTestBase test, bool detailed = true)
+        public static async Task<List<SparkBatchJob>> ListSparkBatchJobsAsync(SparkBatchClient client, bool detailed = true)
         {
             List<SparkBatchJob> batches = new List<SparkBatchJob>();
             int from = 0;
@@ -62,7 +64,7 @@ namespace Azure.Analytics.Synapse.Spark.Tests
             int pageSize = 20;
             do
             {
-                SparkBatchJobCollection page = (await test.SparkBatchClient.GetSparkBatchJobsAsync(detailed: detailed, from: from, size: pageSize)).Value;
+                SparkBatchJobCollection page = (await client.GetSparkBatchJobsAsync(detailed: detailed, from: from, size: pageSize)).Value;
                 currentPageSize = page.Total;
                 from += currentPageSize;
                 batches.AddRange(page.Sessions);
@@ -70,7 +72,7 @@ namespace Azure.Analytics.Synapse.Spark.Tests
             return batches;
         }
 
-        internal static async Task<List<SparkSession>> ListSparkSessionsAsync(this SparkClientTestBase test, bool detailed = true)
+        public static async Task<List<SparkSession>> ListSparkSessionsAsync(SparkSessionClient client, bool detailed = true)
         {
             List<SparkSession> sessions = new List<SparkSession>();
             int from = 0;
@@ -78,7 +80,7 @@ namespace Azure.Analytics.Synapse.Spark.Tests
             int pageSize = 20;
             do
             {
-                SparkSessionCollection page = (await test.SparkSessionClient.GetSparkSessionsAsync(detailed: detailed, from: from, size: pageSize)).Value;
+                SparkSessionCollection page = (await client.GetSparkSessionsAsync(detailed: detailed, from: from, size: pageSize)).Value;
                 currentPageSize = page.Total;
                 from += currentPageSize;
                 sessions.AddRange(page.Sessions);