+++ sdk/communication/Azure.Communication.Sms/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added support to create SmsClient with AzureKeyCredential.
+- Support for creating SmsClient with TokenCredential
 
 
 ## 1.0.0-beta.3 (2020-11-16)